move code screen and fix style
