+++ app/lib/components/codebox.tsx
@@ -26,8 +26,6 @@ interface Props {
 }
 
 export default function CodeBox(props: Props) {
-  const [isLaunchingConsole, setIsLaunchingConsole] = useState<boolean>(false);
-
   const raw = typeof props.code == "string" ? props.code : props.code.raw;
 
   useEffect(() => {
@@ -39,28 +37,6 @@ export default function CodeBox(props: Props) {
       // endregion
     }
   }, []);
-
-  const onQuickLookClicked = (e) => {
-    const setLoadingState = (loading: boolean) => {
-      console.log(loading);
-      setIsLaunchingConsole(loading);
-    };
-
-    setLoadingState(true);
-    quickLook("quicklook", props.app)
-      .then((r) => {
-        setLoadingState(false);
-        PluginSdk.notify("quick look ready !");
-      })
-      .catch((e) => {
-        console.error(e);
-        setLoadingState(false);
-        PluginSdk.notify("compile failed. view console for details.", 2);
-      });
-
-    // ANALYTICS
-    analytics.event_click_quicklook();
-  };
 
   return (
     <>
@@ -79,28 +55,6 @@ export default function CodeBox(props: Props) {
       </CodeWrapper>
 
       {/* FIXME: Need to move that wrapper to parent position  */}
-      <CodeFooterCtaWrapper>
-        {/* copy feature should be inside the code viewer box with copy icon button.*/}
-        {/* <CopyCodeButton onClick={onCopyClicked}>copy code</CopyCodeButton> */}
-
-        <NextStepButton onClick={onQuickLookClicked}>next</NextStepButton>
-        {props.app && (
-          <PreviewButton
-            disabled={isLaunchingConsole}
-            onClick={onQuickLookClicked}
-          >
-            {isLaunchingConsole ? "launching.." : "preview"}
-          </PreviewButton>
-        )}
-
-        {/* for develop */}
-        {/* <PreviewButton
-          disabled={isLaunchingConsole}
-          onClick={onQuickLookClicked}
-        >
-          {isLaunchingConsole ? "launching.." : "preview"}
-        </PreviewButton> */}
-      </CodeFooterCtaWrapper>
     </>
   );
 }
@@ -128,45 +82,15 @@ function PrismCodehighlight(props: { code: string; language: any | Language }) {
   );
 }
 
-const CodeFooterCtaWrapper = styled.div`
-  margin: 0 -8px;
-  padding: 10px 10px 0 10px;
-  display: flex;
-
-  button {
-    margin-right: 8px;
-
-    &:last-child() {
-      margin-right: 0;
-    }
-  }
-`;
-
-const NextStepButton = styled(Button)`
-  ${BlackButton}
-  width: 66.666666%;
-
-  // for reset material-ui button style
-  &:hover {
-    color: #fff;
-    background: #17181a;
-  }
-`;
-
-const PreviewButton = styled(Button)`
-  ${WhtieButton}
-  width: 33.333333%;
-`;
-
 const CodeWrapper = styled.code`
   width: max-content;
   height: auto;
 `;
 
 const CodeInnerWrapper = styled.pre`
-  margin: 0 -8px;
-  padding: 8px;
-  overflow: scroll;
   width: 100%;
   height: 408px;
+  margin: 0 -8px;
+  padding: 8px;
+  /* overflow: scroll; */
 `;
